Code/requirements.txt
appnope==0.1.3
asttokens==2.2.1
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.26.2
backcall==0.2.0
backports.functools-lru-cache==1.6.4
Bottleneck==1.3.5
certifi==2022.12.7
charset-normalizer==3.0.1
debugpy==1.5.1
decorator==5.1.1
executing==1.2.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.15.0
ipython==8.10.0
isodate==0.6.1
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.0.2
jupyter_core==5.1.1
kafka-python==2.0.2
matplotlib-inline==0.1.6
msrest==0.7.1
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
packaging==23.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7
pyzmq==23.2.0
requests==2.28.2
requests-oauthlib==1.3.1
scikit-surprise==1.1.3
scipy==1.10.0
setuptools==65.6.3
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
wheel==0.37.1
zipp==3.13.0
